ported map to microapp, improve state management with the help of rxjs
