+++ packages/dashboard/src/components/doors-app.tsx
@@ -1,30 +1,44 @@
 import { CardActions, Grid } from '@mui/material';
 import { BuildingMap, DoorState } from 'api-client';
 import React from 'react';
-import { DoorCard as DoorCard_, DoorControls } from 'react-components';
+import {
+  DoorCard as BaseDoorCard,
+  DoorCardProps as BaseDoorCardProps,
+  DoorControls,
+} from 'react-components';
 import { DoorMode as RmfDoorMode } from 'rmf-models';
 import { AppRegistry, createMicroApp } from './micro-app';
 import { RmfIngressContext } from './rmf-app';
 
-const DoorCard = React.memo(DoorCard_);
+type DoorCardProps = Omit<BaseDoorCardProps, 'mode'>;
+
+const DoorCard = ({ children, ...otherProps }: DoorCardProps) => {
+  const rmf = React.useContext(RmfIngressContext);
+  const [doorState, setDoorState] = React.useState<DoorState | null>(null);
+  React.useEffect(() => {
+    if (!rmf) {
+      return;
+    }
+    const sub = rmf.getDoorStateObs(otherProps.name).subscribe(setDoorState);
+    return () => sub.unsubscribe();
+  }, [rmf, otherProps.name]);
+
+  return (
+    <BaseDoorCard mode={doorState ? doorState.current_mode.value : -1} {...otherProps}>
+      {children}
+    </BaseDoorCard>
+  );
+};
 
 export const DoorsApp = createMicroApp('Doors', () => {
   const rmf = React.useContext(RmfIngressContext);
   const [buildingMap, setBuildingMap] = React.useState<BuildingMap | null>(null);
-  const [doorStates, setDoorStates] = React.useState<Record<string, DoorState>>({});
 
   React.useEffect(() => {
     if (!rmf) {
       return;
     }
-    const sub = rmf.buildingMapObs.subscribe((newMap) => {
-      for (const door of newMap.levels.flatMap((level) => level.doors)) {
-        rmf.getDoorStateObs(door.name).subscribe((state) => {
-          setDoorStates((prev) => ({ ...prev, [door.name]: state }));
-        });
-      }
-      setBuildingMap(newMap);
-    });
+    const sub = rmf.buildingMapObs.subscribe(setBuildingMap);
     return () => sub.unsubscribe();
   }, [rmf]);
 
@@ -37,9 +51,6 @@ export const DoorsApp = createMicroApp('Doors', () => {
               key={door.name}
               name={door.name}
               level={level.name}
-              mode={
-                doorStates[door.name] === undefined ? -1 : doorStates[door.name].current_mode.value
-              }
               type={door.door_type}
               sx={{ width: 200 }}
             >
