+++ packages/react-components/lib/map/index.tsx
@@ -1,18 +1,14 @@
-export * from './affine-image-overlay';
 export * from './door-marker';
-export * from './doors-overlay';
 export * from './entity-manager';
+export * from './hooks';
 export * from './lift-marker';
-export * from './lifts-overlay';
 export * from './map';
 export * from './robot-marker';
-export * from './robots-overlay';
-export * from './trajectories-overlay';
+export * from './svg-overlay';
 export * from './trajectory';
 export * from './trajectory-marker';
 export * from './trajectory-time-control';
 export * from './utils';
 export * from './waypoint-marker';
-export * from './waypoints-overlay';
+export * from './with-label';
 export * from './workcell-marker';
-export * from './workcells-overlay';
